Maths

Documentation for Algorithms.

Algorithms.area_circle

Algorithms.area_cube

Algorithms.area_heron_triangle

Algorithms.area_hexagon

Algorithms.area_octagon

Algorithms.area_parallelogram

Algorithms.area_rectangle

Algorithms.area_sphere

Algorithms.area_square

Algorithms.area_trapezium

Algorithms.area_triangle

Algorithms.euler_backward_method

Algorithms.euler_forward_method

Algorithms.euler_totient

Algorithms.fibonacci

Algorithms.volume_cone

Algorithms.volume_cube

Algorithms.volume_cuboid

Algorithms.volume_cylinder

Algorithms.volume_parallelepiped

Algorithms.volume_prism

Algorithms.volume_pyramid

Algorithms.volume_sphere

Algorithms.area_circle — Method

area_circle(radius)

Finds area of a circle

Example

area_circle(8) # returns 201.06192982974676
area_circle(-1) # returns DomainError

Algorithms.area_cube — Method

area_cube(side)

Finds area of a cube

Example

area_cube(5) # returns 150
area_cube(-1) # returns DomainError

Algorithms.area_heron_triangle — Method

area_heron_triangle(side1, side2, side3)

Finds area of a heron triangle (a triangle is a triangle where each of the lengths of its sides and its area are expressed in rational numbers.)

Example

area_heron_triangle(2, 12, 13) == 10.790620927453618

area_heron_triangle(0, 0, -1) # returns DomainError
area_heron_triangle(0, -1, 0) # returns DomainError
area_heron_triangle(0, -1, -1) # returns DomainError
area_heron_triangle(-1, 0, 0) # returns DomainError
area_heron_triangle(-1, 0, -1) # returns DomainError
area_heron_triangle(-1, -1, 0) # returns DomainError
area_heron_triangle(-1, -1, -1) # returns DomainError

area_heron_triangle(1, 2, 4) # returns DomainError
area_heron_triangle(1, 6, 4) # returns DomainError
area_heron_triangle(4, 2, 1) # returns DomainError

Algorithms.area_hexagon — Method

area_hexagon(perimeter, apothegm)

Finds area of a hexagon

Example

area_hexagon(5, 4) # returns 10.0
area_hexagon(-1, 0) # return DomainError
area_hexagon(0, -1) # returns DomainError
area_hexagon(-1, -1) # returns DomainError

Algorithms.area_octagon — Method

area_octagon(side)

Finds area of an octagon

Example

area_octagon(6) # returns 173.82337649086284
area_octagon(-1) # returns DomainError

Algorithms.area_parallelogram — Method

area_parallelogram(base, height)

Finds area of a parallelogram

Example

area_parallelogram(4, 50) # returns 200
area_parallelogram(-1, 0) # return DomainError
area_parallelogram(0, -1) # returns DomainError
area_parallelogram(-1, -1) # returns DomainError

Algorithms.area_rectangle — Method

area_rectangle(length, width)

Finds area of a rectangle

Example

area_rectangle(5, 10) # returns 50
area_rectangle(-1, 0) # return DomainError
area_rectangle(0, -1) # returns DomainError
area_rectangle(-1, -1) # returns DomainError

Algorithms.area_sphere — Method

area_sphere(radius)

Finds area of a sphere

Example

area_sphere(8) # returns 804.247719318987
area_sphere(-1) # returns DomainError

Algorithms.area_square — Method

area_square(side)

Finds area of a square

Example

area_square(5) # returns 25
area_square(-1) # returns DomainError

Algorithms.area_trapezium — Method

area_trapezium(Base, base, height)

Finds area of a trapezium

Example

area_trapezium(10, 20, 30) # returns 450.0
area_trapezium(0, 0, -1) # returns DomainError
area_trapezium(0, -1, 0) # returns DomainError
area_trapezium(0, -1, -1) # returns DomainError
area_trapezium(-1, 0, 0) # returns DomainError
area_trapezium(-1, 0, -1) # returns DomainError
area_trapezium(-1, -1, 0) # returns DomainError
DomainError area_trapezium(-1, -1, -1) # returns DomainError

Algorithms.area_triangle — Method

area_triangle(base, height)

Finds area of a triangle

Example

area_triangle(5, 7) # returns 17.5
area_triangle(-1, 0) # return DomainError
area_triangle(0, -1) # returns DomainError
area_triangle(-1, -1) # returns DomainError

Algorithms.euler_backward_method — Method

euler_backward_method()

Find the solution of a first order ordinary differential equation using backward (explicit) euler method

Example

euler_backward_method()

Algorithms.euler_forward_method — Function

euler_forward_method()

Find the solution of a first order ordinary differential equation using forward (implicit) euler method

Example

euler_forward_method((x, t) -> x, 1, (0, 5))[1][end] # returns 143.33937864611195

Algorithms.euler_totient — Method

euler_totient(n)

Find the Euler's totient phi, Euler Totient Function is also known as phi function.

Example

euler_totient(5) # returns 4.0

Algorithms.fibonacci — Method

fibonacci(n)

In mathematics, the Fibonacci sequence is a sequence of integers in which each term is the sum of the two preceding terms.

Example

fibonacci(10) # returns 55

Algorithms.volume_cone — Method

volume_cone(side)

Finds volume of a cone

Example

volume_cone(5, 8) # returns 209.43951023931956
volume_cone(-1, 0) # returns DomainError
volume_cone(0, -1) # returns DomainError
volume_cone(-1, -1) # returns DomainError

Algorithms.volume_cube — Method

volume_cube(side)

Finds volume of a cube

Example

volume_cube(3) # returns 27
volume_cube(-1) # returns DomainError

Algorithms.volume_cuboid — Method

volume_cuboid(side)

Finds volume of a cuboid

Example

volume_cuboid(5, 7, 10) # returns 350
volume_cuboid(-1, 0, 0) # returns DomainError
volume-cuboid(0, -1, -1) # returns DomainError
volume_cuboid(-1, -1, 0) # returns DomainError
volume_cuboid(-1, -1, -1) # returns DomainError

Algorithms.volume_cylinder — Method

volume_cylinder(side)

Finds volume of a cylinder

Example

volume_cylinder(7, 5) # returns 769.6902001294993
volume_cylinder(-1, 0) # returns DomainError
volume_cylinder(0, -1) # returns DomainError
volume_cylinder(-1, -1) # returns DomainError

Algorithms.volume_parallelepiped — Method

volume_parallelepiped(side)

Finds volume of a parallelepiped

Example

volume_parallelepiped(4, 3, 6) # returns 72
volume_parallelepiped(-1, 0, 0) # returns DomainError
volume_parallelepiped(0, -1, 0) # returns DomainError
volume_parallelepiped(0, 0, -1) # returns DomainError
volume_parallelepiped(-1, -1, -1) # returns DomainError

Algorithms.volume_prism — Method

volume_prism(side)

Finds volume of a prism

Example

volume_prism(8, 5) # returns 40
volume_prism(-1, 0) # returns DomainError
volume_prism(0, -1) # returns DomainError
volume_prism(-1, -1) # returns DomainError

Algorithms.volume_pyramid — Method

volume_pyramid(side)

Finds volume of a pyramid

Example

volume_pyramid(3, 5) # returns 5
volume_pyramid(-1, 0) # returns DomainError
volume_pyramid(0, -1) # returns DomainError
volume_pyramid(-1, -1) # returns DomainError

Algorithms.volume_sphere — Method

volume_sphere(side)

Finds volume of a sphere

Example

volume_sphere(3) # returns 113.09733552923254
volume_sphere(-1) # returns DomainError